[thirdweb] Fix waitUntil facilitator parameter not being respected
