+++ .changeset/ten-donuts-fail.md
@@ -1,0 +1,5 @@
+---
+"thirdweb": patch
+---
+
+Fix waitUntil facilitator param not being respected
